+++ deckhouse-controller/cmd/deckhouse-controller/register-go-hooks.go
@@ -61,8 +61,8 @@ import (
 	_ "github.com/deckhouse/deckhouse/modules/038-registry/hooks/orchestrator/users"
 	_ "github.com/deckhouse/deckhouse/modules/039-registry-packages-proxy/hooks/https"
 	_ "github.com/deckhouse/deckhouse/modules/040-control-plane-manager/hooks"
+	_ "github.com/deckhouse/deckhouse/modules/040-control-plane-manager/hooks/autotune"
 	_ "github.com/deckhouse/deckhouse/modules/040-control-plane-manager/hooks/https"
-	_ "github.com/deckhouse/deckhouse/modules/040-control-plane-manager/hooks/resources"
 	_ "github.com/deckhouse/deckhouse/modules/040-control-plane-manager/requirements"
 	_ "github.com/deckhouse/deckhouse/modules/040-node-manager/hooks"
 	_ "github.com/deckhouse/deckhouse/modules/040-node-manager/hooks/migration"
