+++ modules/040-control-plane-manager/hooks/autotune/common_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resources
+package autotune
 
 import (
 	"testing"
