+++ modules/040-control-plane-manager/hooks/autotune/hook_autotune.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resources
+package autotune
 
 import (
 	"context"
